+++ src/main/java/br/com/sismedicina/gestor/repositorios/TecnicoRepositorio.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TecnicoRepositorio extends JpaRepository<Tecnico, Integer> {
+public interface TecnicoRepositorio extends JpaRepository<Tecnico, Long> {
 
 }
